+++ src/runtime_init_memory.js
@@ -11,8 +11,9 @@
 
 // check for full engine support (use string 'subarray' to avoid closure compiler confusion)
 
+function initMemory() {
 #if PTHREADS
-if (!ENVIRONMENT_IS_PTHREAD) {
+  if (ENVIRONMENT_IS_PTHREAD) return;
 #endif // PTHREADS
 
 #if expectToReceiveOnModule('wasmMemory')
@@ -21,7 +22,7 @@ if (!ENVIRONMENT_IS_PTHREAD) {
   } else
 #endif
   {
-    {{{ makeModuleReceiveWithVar('INITIAL_MEMORY', undefined, INITIAL_MEMORY) }}}
+    var INITIAL_MEMORY = {{{ makeModuleReceiveExpr('INITIAL_MEMORY', INITIAL_MEMORY) }}}
 
 #if ASSERTIONS
     assert(INITIAL_MEMORY >= {{{STACK_SIZE}}}, 'INITIAL_MEMORY should be larger than STACK_SIZE, was ' + INITIAL_MEMORY + '! (STACK_SIZE=' + {{{STACK_SIZE}}} + ')');
@@ -53,7 +54,5 @@ if (!ENVIRONMENT_IS_PTHREAD) {
   }
 
   updateMemoryViews();
-#if PTHREADS
 }
-#endif