+++ ChangeLog.md
@@ -33,6 +33,13 @@ See docs/process.md for more on how version tagging works.
   behavior of `Module['preInit']` and compile time callbacks (rather than the
   contrary, as we generally expect an array of functions to be executed left to
   right). (#24012)
+- The standard memory views (HEAP8, HEAP32, etc) are no longer exported by
+  default.  This matches the existing behaviour of `-sSTRICT` and
+  `-sMINIMAL_RUNTIME`.  If you need to access those from outside the module code
+  you can export them by adding them to `-sEXPORTED_RUNTIME_METHODS`.  For
+  example, `-sEXPORTED_RUNTIME_METHODS=HEAP8,HEAPU32` (if you need `HEAP8` and
+  `HEAPU32`). (#24079)
+- libjpeg port updated from 9c to 9f. (#24085)
 
 4.0.6 - 03/26/25
 ----------------